backend/requirements.txt
flask
flask-cors
transformers
torch
accelerate
requests
python-dotenv
SpeechRecognition
pyttsx3
PyAudio
